tests/wpt/mozilla/tests/accessibility-tree/accessibility-update-descendants-of-heading-change.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<h2>
  We really <em>really <strong>really</strong></em> like owls
</h2>

<script>
  setup({ explicit_done: true });
  window.onload = () => {
    test((t) => {
      ServoTestUtils.ensureAccessibilityActive();
      let initial = ServoTestUtils.forceAccessibilityUpdate();

      assert_equals(initial.nodesUpdatedFromDom, 18, "nodesUpdatedFromDom");
      assert_equals(initial.nodesUpdatedFromTree, 18, "nodesUpdatedFromTree");
      assert_equals(initial.nodesInTreeUpdate, 18, "nodesInTreeUpdate");

      document.querySelector("strong").remove();

      let update = ServoTestUtils.forceAccessibilityUpdate();

      // Updated nodes:
      // - em (<strong> removed)
      assert_equals(update.nodesUpdatedFromDom, 1, "nodesUpdatedFromDom");

      // <html>, <body>, <h2>, <em>
      assert_equals(update.nodesUpdatedFromTree, 4, "nodesUpdatedFromTree");

      // Changed nodes:
      // - <h2>, label changed
      // - <em>, children changed
      assert_equals(update.nodesInTreeUpdate, 2, "nodesInTreeUpdate");
    }, "");

    done();
  };
</script>
